requirements.txt
python>=3.10
python-dotenv
langchain_google_genai
langchain_core>=0.2.10
crewai
crewai-tools

